Send presence on new thread
Just in case the other party does not have our presence, clearly we are
ok with them having it if we are starting a thread with them.

Maybe also leak presence on first reply to a new thread from others?
Maybe check first if they have a presence subscription in our roster?